requirements.txt
# All Minnesota - Indigenous-led community meal distribution fundraising site
Django>=5.0
Pillow
django-environ
gunicorn
psycopg2-binary
